refactor: move components functions to component folders
