+++ src/components/_plans/Calendar/helpers/getCalendarMonthValues.ts
@@ -6,8 +6,7 @@ import { DayName } from 'types/components/calendar/dayName.enum';
 import { CALENDAR_WEEK_DAYS } from 'src/constants';
 
 import { betterAt } from 'utils/betterAt';
-
-import { getSplittedToday } from './getSplittedToday';
+import { getSplittedToday } from 'utils/getSplittedToday';
 
 const { PREVIOUS_MONTH_DAY, CURRENT_MONTH_DAY, NEXT_MONTH_DAY } = DayName;
