+++ src/components/_plans/Calendar/helpers/getCalendarYearsValues.ts
@@ -6,8 +6,7 @@ import { YearType } from 'types/components/calendar/yearType.type';
 import { LAST_MIN_START_COUNTING_YEAR, MAX_VISIBLE_YEAR, MIN_VISIBLE_YEAR } from 'src/constants';
 
 import { betterAt } from 'utils/betterAt';
-
-import { getSplittedToday } from './getSplittedToday';
+import { getSplittedToday } from 'utils/getSplittedToday';
 
 const { HIGHLIGHTED, PREVIOUS, NEXT } = YearType;
