+++ src/components/_plans/Events/Events.tsx
@@ -13,7 +13,7 @@ import { AddEventData } from 'types/components/taskbar/events.type';
 import { CreateEvent } from '_plans/CreateEvent/CreateEvent';
 import { EventsList } from '_plans/EventsList/EventsList';
 
-import { getSelectedDateAsString } from 'utils/calendar/getSelectedDateAsString';
+import { getSelectedDateAsString } from 'utils/getSelectedDateAsString';
 
 import { initialEventFormValues } from './data/data';
