+++ src/components/_plans/Events/components/EventsHeader/helpers/changeSelectedDateToDayName.ts
