+++ src/utils/formatCurrentDate/formatCurrentDate.test.ts
@@ -1,4 +1,4 @@
-import { formatCurrentDate } from 'utils/calendar/formatCurrentDate';
+import { formatCurrentDate } from './formatCurrentDate';
 
 test('function should return object with three strings', () => {
   const result = formatCurrentDate(new Date().toLocaleDateString());
