+++ src/utils/getSplittedToday/getSplittedToday.test.ts
