+++ src/utils/getSplittedToday/getSplittedToday.ts
